+++ fancy_core/Rendering/Vulkan/RtAccelerationStructureVk.cpp
@@ -71,7 +71,7 @@ namespace Fancy
 
           const DataFormatInfo& vertexFormatInfo = DataFormatInfo::GetFormatInfo(geoInfo.myVertexFormat);
           const uint vertexComponentSize = vertexFormatInfo.mySizeBytes / vertexFormatInfo.myNumComponents;
-          const uint vertexStride = glm::max(geoInfo.myVertexStride, vertexFormatInfo.mySizeBytes);
+          const uint vertexStride = glm::max(geoInfo.myVertexStride, (uint) vertexFormatInfo.mySizeBytes);
           ASSERT(MathUtil::IsAligned(vertexStride, vertexComponentSize)); // Stride must be a multiple of the component size
 
           VkIndexType indexType = VK_INDEX_TYPE_NONE_KHR;
